Allow passing script arguments to toolbox binary
The current version of the script does not allow commands such as `jetbrains-toolbox --version` to execute correctly, because the arguments to the symlink script are not passed to the binary. This commit fixes that by ensuring the arguments are passed as expected.

Existing users who want the functionality enabled will either have to rerun the download script, or manually replicate this change in their symlinked script.